修复上传文件列表必须配合移除回调函数才能删除的问题

+++ components/upload/upload.jsx
@@ -198,13 +198,11 @@ export default class Upload extends React.Component {
     /* eslint-disable */
     file.status = 'removed';
     /* eslint-enable */
-    if ('onRemove' in this.props) {
+    if ('onRemove' in this.props && this.props.onRemove(file) === false) {
       this.props.onRemove(file);
-      if (this.props.onRemove(file) === false) {
-        return false
-      }
-      this.handleRemove(file);
+      return false
     }
+    this.handleRemove(file);
   }
 
   onChange = (info, updateState = true) => {
